Refactored the logic the filter uses to make it's estimates.
The filter now calls a function, which has a default value that can be overridden. This function accepts the current measurement, and the the current prediction for each value (loc, head, SoG) and returns an estimate. This pattern allows us to smoothly inject whatever logic we like into the filter when it comes to making it's estimates. If we want the filter to make its estimates differently, we simply write the function that describes the deisred decision making, and then pass it to the filter.